+++ moshi-proguard-rule-gen/src/main/kotlin/dev/zacsweers/moshix/proguardgen/MoshiProguardGenSymbolProcessor.kt
@@ -98,6 +98,11 @@ public class MoshiProguardGenSymbolProcessor(private val environment: SymbolProc
               targetConstructorHasDefaults = false,
               targetConstructorParams = emptyList(),
             )
+
+          if (!enableMoshi && nestedSealedClassNames.isEmpty()) {
+            return emptyList()
+          }
+
           environment.logger.info(
             "MOSHI: Writing proguard rules for ${targetType.canonicalName}: $config",
             clazz,